ENG-0000 fix(portal): error page
